+++ installed/include/dexhand_sdk/revo2_hand_can_customed.h
@@ -74,8 +74,6 @@ public:
      */
     GripForce getGripForce() override final;
 
-    bool usesRevo2MotorApi() const override { return true; }
-
 private:
     /**
      * @brief 私有构造函数，用于Connect方法创建实例