+++ src/kuavo_solver/include/kuavo_solver/ankle_solver.h
@@ -3,13 +3,24 @@
 
 #include "Eigen/Dense"
 #include <iostream>
+#include <memory>
+
+namespace kuavo_solver {
+    class Roban_ankle_solver;
+    
+    struct Roban_ankle_solver_deleter {
+        void operator()(Roban_ankle_solver* p) const;
+    };
+}
+
 enum AnkleSolverType {
     ANKLE_SOLVER_TYPE_4GEN = 0,
     ANKLE_SOLVER_TYPE_4GEN_PRO = 1,
     ANKLE_SOLVER_TYPE_5GEN = 2,
     ANKLE_SOLVER_TYPE_S1GEN = 3, 
     ANKLE_SOLVER_TYPE_S2GEN = 4,
-    ANKLE_SOLVER_TYPE_NONE = 5,
+    ANKLE_SOLVER_TYPE_S2GEN_2 = 5,
+    ANKLE_SOLVER_TYPE_NONE = -1,
 };
 class AnkleSolver
 {
@@ -41,10 +52,20 @@ public:
     Eigen::VectorXd motor_to_joint_position_s2_(const Eigen::VectorXd& p);
     Eigen::VectorXd motor_to_joint_velocity_s2_(const Eigen::VectorXd& q, const Eigen::VectorXd& p, const Eigen::VectorXd& v);
     Eigen::VectorXd motor_to_joint_torque_s2_(const Eigen::VectorXd& q, const Eigen::VectorXd& p, const Eigen::VectorXd& c);
+
+    Eigen::VectorXd joint_to_motor_position_s2_2_(const Eigen::VectorXd& q);
+    Eigen::VectorXd joint_to_motor_velocity_s2_2_(const Eigen::VectorXd& q, const Eigen::VectorXd& p, const Eigen::VectorXd& dp);
+    Eigen::VectorXd joint_to_motor_current_s2_2_(const Eigen::VectorXd& q, const Eigen::VectorXd& p, const Eigen::VectorXd& t);
+    Eigen::VectorXd motor_to_joint_position_s2_2_(const Eigen::VectorXd& p);
+    Eigen::VectorXd motor_to_joint_velocity_s2_2_(const Eigen::VectorXd& q, const Eigen::VectorXd& p, const Eigen::VectorXd& v);
+    Eigen::VectorXd motor_to_joint_torque_s2_2_(const Eigen::VectorXd& q, const Eigen::VectorXd& p, const Eigen::VectorXd& c);
     void getconfig(const int ankle_solver_type);
     AnkleSolverType getAnkleSolverType(){
         return  static_cast<AnkleSolverType>(ankle_solver_type_);
     };
+    
+    ~AnkleSolver();
+    
 private:
 
     void applyRollLimitBasedOnPitch(Eigen::VectorXd& joint_q);
@@ -54,6 +75,7 @@ private:
     int ankle_solver_type_ = AnkleSolverType::ANKLE_SOLVER_TYPE_4GEN;
     Eigen::Vector2d ankle_pitch_limits_;
     Eigen::Vector2d ankle_roll_limits_;
+    std::unique_ptr<kuavo_solver::Roban_ankle_solver, kuavo_solver::Roban_ankle_solver_deleter> roban_solver_;
 
 };
 #endif